+++ src/main/model-legacy-extension-api.ts
@@ -40,7 +40,9 @@ export function createLegacyExtensionApi(host: Model, options?: SetDataOptions):
         callAction,
         actionHandler: (actionName: string, ...args: unknown[]) => host.action.has(actionName)
             ? () => callAction(actionName, ...args)
-            : undefined
+            : undefined,
+        getStorageValue: (key: string) => host.storage.getValue(key),
+        setStorageValue: (key: string, value: unknown) => host.storage.setValue(key, value)
     };
     let queryCustomAssertions = {
         ...modelCommonJoraAssertions